refactor to create vep_utils
